v1 which may be vF? idk man
this is actually like v3 but considering i started this all today it's actually v1 bc this is where i'm declaring it completely functional as a standalone. will i be adding more to this later? possibly. idk.

todo: sort grades (descending) and/or other indicator for who'd be eligible for prizing? ideally fix grading system so that average performances get actually average grades and not 3.92 while i'm at it