well, it wasn't a total disaster
possibly my last commit of 2022. depends on how long i continue procrastinating packing.

i now made it export into an excel but, uh. gigo seems to be the case there. not quite sure why that is but i'd like to know. i also tried to balance the grading a bit more and genuinely do not know how to feel on my status for that. idk my head hurts and i need to pack.